Refactor GKE client functions into gke package
GKE client functionality should be co-located with the other GKE methods
in the internal gke package.
