+++ controller/gke-cluster-config-handler.go
@@ -136,7 +136,7 @@ func (h *Handler) OnGkeConfigRemoved(key string, config *gkev1.GKEClusterConfig)
 	if err != nil {
 		return config, err
 	}
-	client, err := utils.GetGKEClient(ctx, cred)
+	client, err := gke.GetGKEClient(ctx, cred)
 	if err != nil {
 		return config, err
 	}
@@ -201,7 +201,7 @@ func (h *Handler) checkAndUpdate(config *gkev1.GKEClusterConfig) (*gkev1.GKEClus
 	if err != nil {
 		return config, err
 	}
-	client, err := utils.GetGKEClient(ctx, cred)
+	client, err := gke.GetGKEClient(ctx, cred)
 	if err != nil {
 		return config, err
 	}
@@ -370,7 +370,7 @@ func (h *Handler) waitForCreationComplete(config *gkev1.GKEClusterConfig) (*gkev
 	if err != nil {
 		return config, err
 	}
-	client, err := utils.GetGKEClient(ctx, cred)
+	client, err := gke.GetGKEClient(ctx, cred)
 	if err != nil {
 		return config, err
 	}
